Exynos5420: Add workaround for ARM errata 799270
This patch adds workaround for the ARM errata 799270 which says
"If the L2 cache logic clock is stopped because of L2 inactivity,
setting or clearing the ACTLR.SMP bit might not be effective. The bit is
modified in the ACTLR, meaning a read of the register returns the
updated value. However the logic that uses that bit retains the previous
value."

BUG=None.
TEST=This workaround is needed only during specific boot, reset or power-down
        sequences, hence, tested with the following patches in this patch-set.

Change-Id: I5b679f8ad56e191334cb16a575188706b15f7fbb
Signed-off-by: Kimoon Kim <kimoon.kim@samsung.com>
Signed-off-by: Akshay Saraswat <akshay.s@samsung.com>
Reviewed-on: https://gerrit-int.chromium.org/37710
Commit-Queue: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Vadim Bendebury <vbendeb@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>