exynos: Mark appropriate MMC devices as non-removable
Both eMMC and SD card devices are served by the mmc driver, which
includes the device structure, which in turn has the 'removable'
property. All mmc devices are presumed removable by the code, but this
is a wrong assumption for eMMC.

One of the ways to tell if the device is removable or not is to
consult the device tree. This change adds some code which will check
the presence of 'samsung,removable' set to zero in the appropriate
device tree node, and if found, will make sure the appropriate device
structure is marked non removable.

BUG=chrome-os-partner:18329
TEST=manual
   . built and booted u-boot code on peach_pit with verified boot
     enabled. Observed the eMMC drive skipped when a removable device
     was looked for.

Change-Id: I6811912a503d883ed78b1ce5b3e7b5f47ffbb725
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/38136
Tested-by: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Simon Glass <sjg@google.com>
Commit-Queue: Vadim Bendebury <vbendeb@google.com>