buildman: Allow make flags to be sepcified for each board
There are a few make options such as BUILD_TAG which can be provided when
building U-Boot. Provide a way for buildman to pass these flags to make
also.

The flags should be in a [make-flags] section and arranged by target name
(the 'target' column in boards.cfg. See the README for more details.

Change-Id: I753583c47dff8b4c1df73174628296fe91f6351b
Signed-off-by: Simon Glass <sjg@chromium.org>