cros: Remove CONFIG_HARDWARE_MAPPED_SPI
Now that U-Boot supports this feature natively, via the 'memory-map' device
tree property, we don't need this feature any more. Removing it simplifies
the code and allows us to implement the feature we really need: partial
reading of the vblock area.

For reference, an example of how to enable the memory map is here:

	spi {
		#address-cells = <1>;
		#size-cells = <0>;
		compatible = "intel,ich9";
		firmware_storage_spi: spi-flash@0 {
			reg = <0>;
			spi-cpol;
			spi-cpha;
			compatible = "winbond,w25q64", "spi-flash";
			memory-map = <0xff800000 0x00800000>;
		};
	};

This will cause all read operations to be performed from the memory-mapped
cache.

This should only affect link/coreboot since it is the only architecture
which uses this feature.

BUG=chromium:205554
TEST=manual
Build and boot on link.

1. Enter recovery mode and see that screens appear
2. Update the EC so that a software sync happens - see from the EC console
that this is taking place.

Change-Id: I5dde2769dcd78476bd1ad2d839fcfd8894dc17a7
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57341
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>