exynos: Reduce compiled size of pinmux driver
Recent changes in proid_is_XXX() and cpu_is_XXX() definitions assign
them to zero if the appropriate SOC code is not compiled in.

This patch makes all SOC specific pinmux initialization code
invocations depend on the appropriate macros, which in turn allows the
compiler to silently drop static functions which are never invoked.

NOTE that there is no actual support for exynos4xxx SOC definitions
yet, so the exynos4 code will not be compiled in.

BUG=chrome-os-partner:18447
TEST=manual
   . with changes to this and other SPL components, the SPL image size
     shrunk below 13K even with pinmux, gpio and serial driver
     compiled in.
   . built and booted peach_pit and snow

Change-Id: I240135fee391de496419df46cbd5d7caf937d322
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit-int.chromium.org/38489
Tested-by: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Simon Glass <sjg@google.com>
Commit-Queue: Vadim Bendebury <vbendeb@google.com>