dwmmc: Fix FIFO_DEPTH calculation
To get the FIFO depth the DWMMC driver reads the default value of
RX_WMARK in the FIFOTH register. Based on the FIFO depth, we calculate
the value of RX_WMARK i.e  (FIFO_DEPTH/2 - 1) and this value was written
back to the FIFOTH register.
In the current code instead of storing the actual FIFO depth, we were
everytime reading the FIFOTH register and extracting the value from
RX_WMARK bits, which is not the correct value of FIFO_DEPTH.
Hence the mmc_init function used to fail after every 3 times, causing
the mmc devices to stop responding properly

BUG=chrome-os-partner:19569
TEST=Tested multiple times on SMDK5420 board. On u-boot prompt following
can be seen
	SMDK5420 # mmc rescan; echo $?
	0
	SMDK5420 # mmc rescan; echo $?
	0
	SMDK5420 # mmc rescan; echo $?
	0
	SMDK5420 # mmc rescan; echo $?
	0
	SMDK5420 # mmc rescan; echo $?
	0
	SMDK5420 # mmc rescan; echo $?
	0

Change-Id: I4dc6d8fb0a40f9ff7c59daf653890db6e35d0fd5
Signed-off-by: Rajeshwari Shinde <rajeshwari.s@samsung.com>
Signed-off-by: Hatim Ali <hatim.rv@samsung.com>
Reviewed-on: https://gerrit-int.chromium.org/38543
Reviewed-by: Alim Akhtar <alim.akhtar@samsung.com>
Commit-Queue: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Vadim Bendebury <vbendeb@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>