Add CONFIG_CMDLINE to allow removal of all commands
A large chunk of the U-Boot code is the command line. For some applications
this is not needed, since the boot can be controlled by a board-specific
hard-coded boot procedure. This saves a significant amount of space.

This makes U-Boot proper look a little more like SPL in terms of size.

Signed-off-by: Simon Glass <sjg@chromium.org>
BUG=chromium:247392
TEST=manual
Build and boot on link and sandbox and see that verified boot works on link.

Build with future changes to add the vboot_stop execution, and see that
verified boot works as before, but image size is now >100KB smaller.

Change-Id: I284f988f311d827a24c0d41e1e0a576d7d753374
Reviewed-on: https://gerrit.chromium.org/gerrit/57822
Commit-Queue: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>