exynos: Add new spl_machine_param for two more board ID GPIOs
We now support an addition two GPIOs for a total of 4. Add a new 'R' machine
parameter to handle this.

BUG=chrome-os-partner:19179
BRANCH=none
TEST=manual
Run new cros_bundle_firmware with -v3, see that it outputs the GPIOs that it
stores.

  Board ID GPIOs: bit0=76, bit1=75
  Board ID GPIOs: bit2=73, bit3=73

Currently SPL does not use this information, so the U-Boot part of this will
be tested when serial and board ID is implemented in SPL.

Change-Id: Ie40c740df73c34ea3a5c874f5338d242f36c69f4
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57795