exynos: config: Enable Thumb builds for Exynos5
This makes U-Boot smaller with no obvious ill effects on the 4.7 toolchain
we use. We enabled it for snow previously, so should do so for pit,
preferably now so that we can catch any problems.

This requires a slight change to one of the exynos macros which assumes
we are running in ARM state.

BUG=chrome-os-partner:18447
BRANCH=none
TEST=manual
Build and boot on pit and snow. No problems noticed.

Size for pit drops by about 24% for U-Boot and 30% for SPL.

This command builds U-Boot for pit:
   -V : verified boot enabled
   -z : display image sizes
   -w : don't write image to board
   -i : incremental build

For ARM we get:

$ crosfw  -Vzwi
   text	   data	    bss	    dec	    hex	filename
 550949	  19060	 303676	 873685	  d54d5	b/chromeos_peach/u-boot
   7328	    368	      0	   7696	   1e10	b/chromeos_peach/spl/u-boot-spl

With this change (i.e. Thumb code):
   text	   data	    bss	    dec	    hex	filename
 419159	  19060	 303700	 741919	  b521f	b/chromeos_peach/u-boot
   5128	    368	      0	   5496	   1578	b/chromeos_peach/spl/u-boot-spl

Tested both with and without the Thumb mode change on pit. The Thumb code
looks like this, which I believe is the same as the intended result.

 202472c:	23d3      	movs	r3, #211	; 0xd3
 202472e:	f383 8100 	msr	CPSR_c, r3

and for ARM state:

 2024ed0:	e3a030d3 	mov	r3, #211	; 0xd3
 2024ed4:	e121f003 	msr	CPSR_c, r3

Change-Id: I5515da2e4970275c2eafac51e81741268fccaeb1
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57460
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>