video: Add driver for Analogix ANX1120 eDP->LVDS bridge chip
This chip provides a means of driving an LVDS display from an eDP output.
The U-Boot driver sets this up ready for use.

This driver comes from the smdk5420.c code. The initialization data has
been obtained from access to an application note for the chip.

BUG=chrome-os-partner:18329
TEST=manual
Boot on pit and see that the display still runs (i.e. we have not broken
parade operation).

Change-Id: Ie90b7dbf9973e36c5a68fc9acf47799a267d8093
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58394
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>