arm: Set cache correctly when DRAM bank ends at 0xffffffff
On systems where a DRAM bank ends at the very top of the address space, the
cache setup code is currently skipping the final bank. Since U-Boot is
relocated to the top of RAM this effectively means the cache is off for
U-Boot.

Fix calculations to make sure that the top memory bank is not ignored.

BUG=chrome-os-partner:19731
TEST=manual
   . boot on 3.5GB pit, type 'print' repeatedly and see that the
     display scrolls quickly with no lag now.

Change-Id: Ifd4a904d9513b1db31fa4ef48f61b619fcda5c58
Signed-off-by: Simon Glass <sjg@chromium.org>
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>