arm: Display unrelocated address for PC/LR in exceptions
When an exception is reported, it is useful to display the unrelocated PC
and LR so that these can be looked up in System.map. Add code to print
this information out.

BUG=none
TEST=manual
Cause an unaligned exception and see that the crash dump now reports an
address within the code and can be looked up in the assembly dump.
Change-Id: Ie59811e0d6658ac017b9e744f74b24d5e27b9bec
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57516
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>