Refactor command macros so they can compile to nothing
Commands are declared entirely within a macro, like this:

U_BOOT_CMD(...);

Sub-commands are not the same. At present the U_BOOT_CMD_MKENT() macro is
used within an array declaration, but the array itself is not handled by
macros. We have this:

static cmd_tbl_t table[] = {
	U_BOOT_CMD_MKENT(...),
	U_BOOT_CMD_MKENT(...),
};

There are two problems with this. The first is that it requires knowledge
of the command type (a minor issue). The second is that it is not possible
to make sub-commands compile to nothing if we want to.

Introduce a new syntax which matches U_BOOT_CMD() as follows:

U_BOOT_SUBCMD_START
	U_BOOT_CMD_MKENT(...)
	U_BOOT_CMD_MKENT(...)
U_BOOT_SUBCMD_END

Note that the U_BOOT_CMD_MKENT() lines no longer have a comma at the end.
It would be possible to put a semicolon after U_BOOT_SUBCMD_END, but it
seems syntactically odd to do so.

BUG=chromium:247392
TEST=manual
Build and boot on link and sandbox. Type 'help' and see all the comamnds
Run the verified boot flow on link.
Change-Id: I3b01ead1d7f7fe506be47acc12dd552baa5b612f
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57821