Add an info word to commands
In cmd_bootm.c the function pointer is misused as a flags word. It is better
to add a flag word to each command, so that bootm can avoid this hack.

To avoid increasing the size of the U-Boot binary, we can re-use the existing
'repeatable' flag, which only needs a single bit.

BUG=chromium:247392
TEST=manual
Build and boot on link and sandbox. Type 'help' and see all the comamnds.
Run the verified boot flow on link. Try bootm subcommands and see that they
are executed correctly.

Build on pit with both options:

This builds a small image:

crosfw -Vwz --small
Configuring for chromeos_peach board...
   text	   data	    bss	    dec	    hex	filename
 211015	   8480	  83948	 303443	  4a153	b/chromeos_peach/u-boot
   5326	    492	      0	   5818	   16ba	b/chromeos_peach/spl/u-boot-spl

This builds a normal image:

crosfw -Vwz
Configuring for chromeos_peach board...
   text	   data	    bss	    dec	    hex	filename
 405500	  17712	 303344	 726556	  b161c	b/chromeos_peach/u-boot
   5326	    492	      0	   5818	   16ba	b/chromeos_peach/spl/u-boot-spl

So the size reduction on pit is 190KB at present (about 48%).

Change-Id: I899f7affdd7da135a5452f4447a3b4ea0a39284f
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57820