cros: Allow EC to be in a separate section
A feature was added to remove the need to load the EC image with U-Boot.
This was intended to improve boot time. It had a noticeable affect on
link with its very slow SPI flash (about 70ms I think) but not on
snow (7ms).

The feature was built on a hack for software sync, since there was not
time to change the signers or much else at that stage of the project.

The hack works like this: The U-Boot section contains the EC, but the
signer is not told about it so only signs U-Boot and the device tree.
Then on boot, vboot therfore loads only U-Boot and the device tree.
Later, when software sync runs it can retrieve the EC from flash and
check that the hash is correct.

Unfortunately this hack does not work with compression, since in that
case we must load the whole region in order to decompress the EC. The
current code simply fails and it seems bad to add further to this hack.

But we can implement this properly just by placing the EC images in their
own section and putting an expected hash in the device tree.

This fix is pretty small, so implement this, removing the old hack in the
process.

BUG=chrome-os-partner:17606
TEST=manual
Add an EC region to the flashmap:

		rw-a-ec {
			label = "ec-rw-a";
			reg = <0x002d0000 0x0001ffc0>;
			type = "blob ecbin";
			add-hash;
		};

'vboot_test fmap' now shows the correct EC details.

Build and boot on link, see that normal boot flow works. Flash the EC
with a different image. Run normal flow again and see that software
sync completes successfully.

Change-Id: If2bb1ad1bf7b18e6ec204dbe4e5120599ddfe44b
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57342
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>