cros: Remove last bits of the physical recovery switch
This code is a hang-over from the old recovery switch, even though most of
it has been removed. We must remove this part also since otherwise we
get an au uninitialized variable and vboot may or may not decide to go
into recovery mode.

BUG=chrome-os-partner:18329
TEST=manual
(Testing on link since it seems to have no other problems)
Boot on link and see that it no longer goes into recovery mode on each
boot.

Change-Id: Ia6ae67bccb828859df380104864678e5415a7dcf
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57338
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>