gen: bootstage: Allow forcing a bootstage report on a silent console
Add a 'bootstage-force-report' config option to the device tree, to produce
a bootstage report even if the console is silenced. This provides an easy
way to time the boot without enabling the console (which slows the boot
down considerably on some platforms).

BUG=none
TEST=manual
$ crosfw -b coreboot -VCS -a '--add-config-int force-bootstage-report 1'
See that we get a bootstage report now, but the console is otherwise
silent.

Change-Id: I9794434d4c6c8af25f6cda6e92acdcc16e406f0d
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57453
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>