gen: Display pre-relocation addresses in initcall
The addresses currently displayed are not very useful without knowing the
relocation offset. Ideally we would have a list of function names, but
for the moment subtracting the relocation offset lets us look up the
address in System.map.

BUG=none
TEST=manual
Adjust one of the post-relocation initcalls to return 1 (indicating
failing), boot on link and see that we now get a useful address.

Change-Id: I54df5fadf336dba8a2b35e08533b09a6ae682461
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57636
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>