gen: console: Add silent-config device tree configuration
This option causes the console to be silent by default. It is then
possible to chose this at run-time using the device tree.

BUG=chrome-os-partner:19927
TEST=manual
Build and boot on link with and without the -S silent flag, and see that
it is honoured:

$ crosfw -b coreboot -V
(console output appears on boot)

$ crosfw -b coreboot -VS
(no console output appears on boot)

Change-Id: I446ab8468746de5261ab2b95f321897ccf7dffbb
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57454
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>