hack: Exynos: DW_MMC: Add delay in interrupt status polling
Now cache is enabled and processing is faster but readl/writel require
same time as before. After "command send" we poll for interrupt
reception which becomes erroneous because of this read time lag. This
is the cause for incorrect values provided by the "mmcinfo"
command. Hence, adding a delay of 1 usec after every read of raw
interrupt status register.

The problem with this fix is that the delay must be happening in this
very place, putting it in the beginning or the end of the do{} while()
section does not fix the issue. Submitting this as a temp workaround,
the root cause will have to be found.

Also, the delay of 100 usecs provided at the end of every "command
send" created lot of delay and is not required.  Hence, removing this
delay.

BRANCH=none
BUG=chrome-os-partner:19816
TEST=Tested mmc read/write and mmcinfo over SCP SMDK 4 GB
	    and peach_pit 2 GB and saw appropriate results.
Change-Id: I83ca797ec47b941cff1b395ff3bfc91abfc1db9e
Signed-off-by: Akshay Saraswat <akshay.s@samsung.com>
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57492
Reviewed-by: Simon Glass <sjg@chromium.org>