cros: dts: Keep the EC regions wiped for platforms with no EC
Since butterfly, stout and parrot will still use this tree, and do not have
an EC, the EC addition that was recently added needs to move. The new location
is only included by link, which is the only platform with an EC. This fixes
a problem where the other platforms expect to find an EC to put into the
image, and cannot.

TBR=build break
BUG=none
TEST=manual
Build butterfly chromeos-u-boot locally with cros_workon, see that the
correct dts file is installed. Build chromeos-bootimage and see that it
succeeds and does not require an EC file.

Change-Id: Ic46f6edcf7bfba3f69dd16d3a769f2f7e8759bb4
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57550
Reviewed-by: David Hendricks <dhendrix@chromium.org>