exynos: fdt: Switch 5250 i2c periphs to s3c2440-i2c
On the 5250, these i2c banks do not work as exynos5-hsi2c.  In fact, the
registers given do not match those in the reference manual for high
speed I2C.  Since they are different between the 5250 and 5420, move
the nodes to exynos54xx.dtsi and exynos52xx.dtsi.  exynos52xx.dtsi was
newly created, so it was added to the board device tree includes.

BRANCH=none
BUG=none
TEST=manual. Boot spring to u-boot, verify i2c peripherals, such as
cros-ec, functional.

Change-Id: Ibce61493c9e0d3a1e8b462aeccd5c28f339d17d8
Signed-off-by: Michael Pratt <mpratt@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57845
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>