exynos: fdt: Update GPIO numbers
The new Exynos GPIO scheme includes an additional bank of GPIOs (C4)
that was not present in the previous branch of u-boot.  This bank starts
at GPIO 160, so GPIOs in the old format >= 160 need to be incremented
by 8 to be in the correct bank.

snow's cros-ec gpio should match spring's.

BRANCH=none
BUG=none
TEST=manual.  On spring, cros-ec communication is functional, including
the keyboard.

Change-Id: I563d77ea2bb3d9798ca7a30c8023dc539f331732
Signed-off-by: Michael Pratt <mpratt@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57850
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>