power: tps65090_init return -ENODEV if i2c bus not found
If the i2c bus of the tps parent node cannot be found, return -ENODEV
rather than a generic -1 error.  Also fix an uninitialized dereference
of the pmic struct.

This allows Exynos boards to fallback quietly if the tps node is not a
direct i2c descendant in the device tree.

BRANCH=none
BUG=none
TEST=manual. Spring successfully boots to U-Boot, without failing
tps60509 initialization.

Change-Id: I14ff4e8c594eb53b3abd08a6760d73606b24dee3
Signed-off-by: Michael Pratt <mpratt@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57852
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>