Exynos: DWMMC: Fix for incorrect EXT CSD read
For a reason not completely understood, the read eMMC CSD data values
are corrupted, even though the d-cache for the stack allocated receive
buffer is invalidated when the eMMC read command is issued.

One of attempted explanations was that there must be a prefetch
corrupting dirtying the receive buffer before actual data arrives. But
it did not hold water - increasing buffer padding both at the bottom
and at the top to 64 bytes did not prevent the issue from happening.

The current running theory is that some speculative execution is going
on and data could be prefetched because of that. Also seems a long
shot. This issue needs to be more thoroughly investigated.

This fix, however, seems to be addressing the problem, so we submit it
to be able to move forward.

BRANCH=none
BUG=chrome-os-partner:19816
TEST=Tested mmc read/write and mmcinfo over SCP SMDK 4 GB
        and peach_pit 2 GB and saw appropriate results.

Change-Id: Ic6af31f22c490e50c1a3e350d3fdb9d8c37072d0
Signed-off-by: Sunil Mazhavanchery <Sunilm@samsung.com>
Signed-off-by: Akshay Saraswat <akshay.s@samsung.com>
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58272
Reviewed-by: Simon Glass <sjg@chromium.org>