drop: usb: increase power on to power good delay
The USB standard (USB 2.0 7.1.7.3) mentioned that the delay between the
power ok on a hub port and the USB device attach detected should be at
most 100ms.
By the way, I have measured :
- about 240ms on Patriot Supersonic Xt 8GB USB keys (black with blue XT).
- about 370ms on Sandisk SDCZP-4096.
- from 450 to 505ms on 64GB USB 3.0 Corsair Survivor Stealth.
So let's extend that delay.

This is a hack, we might eventually want to do something more
sophisticated. If we end up using U-Boot for that long...

Change-Id: I6c0e9cf7f211d0352d0c9e92583e227e4b10c3bc
Signed-off-by: Vincent Palatin <vpalatin@chromium.org>
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/63501