cros: dts: Add bad-wake-gpio for exynos5420-peach_pit
An invalid wakeup happens when the state of a certain GPIO is high
during wakeup.  This is used to detect a malicious user who might be
trying to mess with the "is sleeping" GPIO.  We detect this in U-Boot
upon wakeup and reboot if we find this attack.

On pit this GPIO is used to prevent the TPM from getting reset across
sleep / wake.

BRANCH=none
CQ-DEPEND=CL:Ie01e1578e136809577127695504109625c5a98f6
BUG=chrome-os-partner:20172
TEST=Run the platform_TpmSleepTestServer autotest on pit

Change-Id: Ia70827284fded79c988c1027a4da7531c6f7f590
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58772
Reviewed-by: Simon Glass <sjg@chromium.org>