exynos: Display warning when rebooting after a watchdog reset
At present the kernel is dying with a watchdog reset, so it seems useful
to display a warning in this case.

BUG=chrome-os-partner:20181
TEST=manual
Boot my kernel on pit with earlyprintk and see that it resets.
U-Boot then prints watchdog warning on next reset.
Reset manually with servo and see that the message is not printed.

Change-Id: I698e875def28bfd6f03d237a29b7b84e12f8fd9d
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58619
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>