fthread: Cooperative multi-threading library
This patch adds the machine-independent files for a cooperative
multi-threading library in U-Boot called fthread, which is a heavily
watered down rewrite of GNU Portable Threads.  Functions that provide
machine-specific context manipulation functions must be provided before
the library can be used on that machine.  Please see doc/README.fthread
for the motivations behind this library as well as porting and usage
instructions.

BRANCH=none
BUG=chrome-os-partner:19927
TEST=manual
Built U-Boot with FTHREAD_CONFIG and FTHREAD_CMD_CONFIG, ran 'fthread
test' and saw that it reported a successful test run.  The test command
will be added in a later commit.

Change-Id: I5fe63ae384804362e8c4424ee93b1e3c31c6da29
Signed-off-by: Chirantan Ekbote <ekbotec@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58765
Reviewed-by: Simon Glass <sjg@chromium.org>