exynos: do not use automatic variable array in exynos_dp
The compiler generated preamble code uses 32 bit access to initialize
this array when it is devined as an automatic variable. Depending on
other program's components the 32 bit value the preamble code reads
could be placed at an unaligned address, in which case the traget just
dies. It does not even report an exception, because exception
processing is not yet enabled.

The fix is to move the array out of the function, where it does not
have to be initialized.

BRANCH=none
BUG=chrome-os-partner:20228, chrome-os-partner:20253
TEST=manual
  both peach_pit and snow now boot.

Change-Id: Ic18dec6825773749eb91369a718627abc9bab053
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58790
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>