cros: Limit the size of the vblock read from SPI flash
At present there is no vboot API to read the GBB and vblock from SPI flash -
this must be done by the vboot_twostop code. This feature will involve a
non-trivial amount of work, so in the meantime, hard-code the number of
bytes of the vblock region which must be read.

BUG=chromium:205554
TEST=manual
Build and boot on link and see that a normal boot succeeds.

See that it now boots faster.

Change-Id: Ic25f7998d338c55e03012b74a402841f1b234652
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/59511
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>