gen: exynos: allocate memory to store color conversion map
When displaying 8 bit BMP images on panels with larger number of bits
per pixel there is a need to map 8 bit colors into a larger color
space.

The map is created on the fly when the BMP image is parsed, some
memory is required to store the map.

This change brings Exynos platforms into the fold: Allocate memory to
store the map at initialization, store the pointer in the panel_info
structure and use it when displaying 8 bit BMP images.

Also, do not silently use the NULL pointer, complain about it and
leave.

BRANCH=none
BUG=none
TEST=manual
   . the

     'vbexport_test image'

     command displays correctly shaped and colored images on pit and
     snow.

Change-Id: I7001dc701b03b90c94a5976e9e117272d53283bd
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/59585