gen: Implement console recording
Serial console output takes fair amount of time, it should be
suppressed to speed up the boot process.

In case of boot problems, however, it is very important to be able to
examine console output for clues as of what went wrong.

Console recording feature implemented in this patch allows just that.
While the system is booting, all console output is saved in a buffer.
If the system is configured for silent console, nothing is printed.

In case of boot failure the dump_console() function can be called. If
console output was disabled - this function enables it and then sends
all recorded information to the console.

The same buffer where console output is recorded could be passed to
the kernel to allow examining firmware console from user space.

Note that this implementation stops recording if and when the buffer
limit is reached without giving any indication that part of the output
has been dropped. If necessary, some error message could be put in the
end of the buffer to highlight the problem.

BRANCH=none
BUG=chrome-os-partner:10808
TEST=manual
  . tested with other patches applied, observed console output printed
    in case of failing boot on the otherwise silent console.

Change-Id: Id7d0aa45f65099d9cc9a6aae35425651d71e1f06
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/59884