cros: Convert boot_device to use linker lists
Before adding a new boot device for sandbox, let's change the existing code
over to use linker lists. This is a much cleaner way of doing this as it
avoids the manual probing.

BUG=chromium:244019
BRANCH=none
TEST=manual
Build and boot normal and recovery mode on snow and link (tests mmc, usb,
scsi). IDE is not tested as it is only used on Alex and I don't have one to
hand.

Change-Id: I504272535e6b88ad03973ba16621d7ddcec6bd7c
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61841
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>