fthread: Fix unaligned access in demo command
This patch fixes an aligned access error that occurs on ARM platforms
when strings are declared in the following format:

    char foo[] = "foo";

While this declaration is legal, it causes the compiler to generate
native loads from the literal string and native stores to the buffer,
and the literal string alignment cannot be controlled. If it is
misaligned, then the core will throw a data abort exception.  For a more
detailed explanation, please see doc/README.arm-unaligned-accesses.

BRANCH=none
BUG=none
TEST=manual
Built U-Boot for ARM, ran "fthread demo" and saw that it didn't give an
unaligned access error.

Change-Id: Ib39989e0fa7761108b6f8138b4b9c28c2e5d3651
Signed-off-by: Chirantan Ekbote <ekbotec@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/60017
Reviewed-by: Simon Glass <sjg@chromium.org>