exynos: dts: Adjust the cros-ec node to work with software sync
The flash description for the Chromium OS EC used to be in its own node.
Now it has been moved to the main EC node, so has to become part of this
(upstream) code.

Snow does not use SPI to talk to the EC, so remove that node - since
otherwise we will find the EC twice. When the flash description was
in its own node this worked, but now it does not.

Make the necessary changes.

BUG=chrome-os-partner:20608
TEST=manual
See that software sync now find the EC and starts correctly on snow. It
does not complete due to an unrelated problem.

> cros_test swsync rw
SF: Detected W25Q32DW with page size 4 KiB, total 32 MiB
Flashing RW EC image: offset=d9000
firmware_storage_spi.c: read_spi: offset=0xd9000, count=0xb5d4
read,
Software sync failed with error -1
SMDK5250 #

(The failure is an already-existing condition in the i2c driver which
I have not yet diagnosed)

Change-Id: I920d8eab5d99e05b350ea21d1fc5b13d774569f9
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/60028