gen: arm: Pass console buffer to linux through device tree
Recent introduction of the console recording ability makes it possible
to pass the console buffer record to the kernel to make console log
available after Linux boots up.

console_get_record() API is being introduced to provide access to the
console record. The device tree passed to the kernel needs to be
padded to allow adding the recording. Another nuance of this change is
that we want the recording to be added in the very end, right before
starting the downloaded kernel. To make this possible we need to
prevent resizing the device tree which removes the padding. Note that
resizing does not save any memory anyways.

BRANCH=none
BUG=chrome-os-partner:10808
TEST=manual
  . on the booted system check the device tree:

   localhost ~ # ls -l /proc/device-tree/chosen/ap-console-buffer
   -r--r--r-- 1 root root 12327 Jun 19 19:50 /proc/device-tree/chosen/ap-console-buffer

   verify that the contents of this file is indeed console output:

   localhost ~ # tail !$
    tail /proc/device-tree/chosen/ap-console-buffer
       Verifying Hash Integrity ... sha1+ OK
       Booting using the fdt blob at 0x203b4b78
       XIP Kernel Image (no loading done) ... OK
    OK
       Loading Device Tree to 3ffe6000, end 3ffffea1 ... OK
    nodeoffset = 180
    boot_kernel.c: ft_board_setup: Completed setting up fdt information for kerrnel

    Starting kernel ...

Change-Id: I83a7f28fcceb44929ab6bd8ac8b2773afd1c4089
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/60615
Reviewed-by: Simon Glass <sjg@chromium.org>