mmc: exynos: Add simple dwmmc driver
Add a simplified dwmmc driver intended for use in SPL.  It is meant to be
used in environments where the entire MMC framework cannot be included,
due to code size or excessive dependencies.

This driver is largely based on the full driver used in U-Boot, with support
for reading and writing data blocks not included.

This driver can be used to send commands that do not require a data transfer,
such as configuring power-on write protection.

When CONFIG_SPL_MMC_SUPPORT is defined, CONFIG_SPL_MMC_SIMPLE can be defined
to build this driver in place of the standard libmmc.

BRANCH=none
BUG=chromium:254233
TEST=manual. With boot write protection patches, boot to U-Boot, verify boot
partition write protection enabled.

Change-Id: I7009e2b09b5bb25244b69d3015e086b1fc135299
Signed-off-by: Michael Pratt <mpratt@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61205
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>